docker/document-processor/requirements.txt
# Serveur web
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Traitement de documents
python-multipart==0.0.6
aiofiles==23.2.1

# Extraction de texte
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2

# OCR
pytesseract==0.3.10
Pillow==10.1.0

# Traitement d'images
opencv-python-headless==4.8.1.78
numpy==1.24.3

# IA/ML pour classification et extraction
scikit-learn==1.3.2
nltk==3.8.1
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2

# Utilitaires
requests==2.31.0
loguru==0.7.2
pydantic==2.5.0
python-magic==0.4.27
chardet==5.2.0

# Base de données pour cache
redis==5.0.1

# Monitoring
prometheus-client==0.19.0